refactor(tls): rename reserve → layout allocator & restructure modules;
