--- src/tls/dtv.rs
@@ -1,3 +1,0 @@
-use std::sync::atomic::AtomicUsize;
-
-static GLOBAL_GENERATION_COUNTER: AtomicUsize = AtomicUsize::new(0);
